Add files via upload
Let's train Huggy the Dog 🐶 to fetch a stick

- **The State Space**: what Huggy "perceives."
Huggy doesn't "see" his environment. Instead, we provide him information about the environment:

The target (stick) position
The relative position between himself and the target
The orientation of his legs.
Given all this information, Huggy can decide which action to take next to fulfill his goal.

- **The Action Space**: what moves Huggy can do
- **The Reward Function** : The reward function is designed so that Huggy will fulfill his goal : fetch the stick.

Remember that one of the foundations of Reinforcement Learning is the reward hypothesis: a goal can be described as the maximization of the expected cumulative reward.

Here, our goal is that Huggy goes towards the stick but without spinning too much. Hence, our reward function must translate this goal.